feat: implement war/diplomacy/economy engines, auxiliary entities, and full frontend UI overhaul
Backend:
- War system: BattleEngine combat formulas, BattleService retreat/capture, BattleTrigger pre/post hooks, WarFormula
- Diplomacy: full lifecycle (propose/accept/reject non-aggression, war declaration, ceasefire)
- Economy: public entry points for event-driven income/supply/semi-annual processing
- Turn engine: traffic update, strategic command limit reset per turn
- AI: expanded GeneralAI/NationAI decision trees
- Events: compound actions, economy delegation, delete_self
- New entities: Auction, Betting, Board, Tournament, Vote, WorldHistory, RankData, GeneralRecord, etc.
- DB migrations V4-V7: auxiliary tables, world name, world_id on turns, general fields
- LastTurn refactored with addTermStack/getTermStack, ConstraintChain added
- World management: delete, reset, rename endpoints
- Comprehensive test coverage for all new systems

Frontend:
- Major UI overhaul across all game pages (city, nation, diplomacy, battle, etc.)
- New pages: battle-center, dynasty, emperor/detail, generals/[id], npc-list, processing, superior, vote/[id]
- Updated components, API client, stores, and types for full backend parity

+++ backend/src/main/kotlin/com/opensam/command/general/DomesticCommand.kt
@@ -47,7 +47,7 @@ abstract class DomesticCommand(
     override suspend fun run(rng: Random): CommandResult {
         val date = formatDate()
         val stat = getStat()
-        val trust = maxOf(50, city?.trust ?: 50)
+        val trust = maxOf(50F, city?.trust ?: 50F).toDouble()
 
         var score = (stat * (trust / 100.0) * (0.8 + rng.nextDouble() * 0.4)).toInt()
         score = maxOf(1, score)
