+++ backend/src/main/kotlin/com/opensam/command/general/che_기술연구.kt
@@ -34,7 +34,7 @@ class che_기술연구(general: General, env: CommandEnv, arg: Map<String, Any>?
     override suspend fun run(rng: Random): CommandResult {
         val date = formatDate()
         val intel = general.intel.toInt()
-        val trust = maxOf(50, city?.trust ?: 50)
+        val trust = maxOf(50F, city?.trust ?: 50F).toDouble()
 
         var score = (intel * (trust / 100.0) * (0.8 + rng.nextDouble() * 0.4)).toInt()
         score = maxOf(1, score)
